+++ packages/headless-styles/src/index.ts
@@ -4,7 +4,6 @@ export { getJSBadgeProps, muiLabelOverride } from './components/Badge/badgeJS'
 export {
   getButtonProps,
   getDangerButtonProps,
-  getIconButtonProps,
 } from './components/Button/buttonCSS'
 export {
   getJSButtonProps,
@@ -22,6 +21,8 @@ export { getJSFormLabelProps } from './components/FormLabel/formLabelJS'
 
 export { getIconProps } from './components/Icon/iconCSS'
 export { getJSIconProps } from './components/Icon/iconJS'
+
+export { getIconButtonProps } from './components/Button/iconButtonCSS'
 
 export { getProgressProps } from './components/Progress/progressCSS'
 export { getJSProgressProps, muiReset } from './components/Progress/progressJS'
